+++ test/babashka/classpath_test.clj
@@ -16,11 +16,11 @@
   ;; a spec.alpha or a tools.deps stand-in on the classpath cannot load in
   ;; bb, so the bundled sources serve those namespaces before the classpath
   (let [dir (fs/create-temp-dir)]
-    (doseq [path ["clojure/spec/alpha.clj" "clojure/tools/deps/util/maven.clj"]]
+    (doseq [path ["clojure/spec/alpha.clj" "clojure/tools/deps.clj" "clojure/tools/deps/util/maven.clj"]]
       (fs/create-dirs (fs/parent (fs/file dir path)))
       (spit (fs/file dir path) "(throw (Exception. \"classpath version loaded\"))"))
     (is (true? (bb nil "--prn" "--classpath" (str dir)
-                   "(require '[clojure.spec.alpha :as s] 'clojure.tools.deps.util.maven) (s/valid? int? 1)")))))
+                   "(require '[clojure.spec.alpha :as s] 'clojure.tools.deps 'clojure.tools.deps.util.maven) (s/valid? int? 1)")))))
 
 (deftest classpath-test
   (is (= :my-script/bb